Fixed a bug with submaster values being treated as ints and not floats as it should be. This caused the submasters to set 100% normally, and all other values as 0%.

+++ main/src/app.h
@@ -124,7 +124,7 @@ class App : public QMainWindow
   int hasSubmaster(t_channel);
   void resetSubmasters();
   void setSubmasterValue(t_channel, int);
-  int submasterValue(t_channel);
+  float submasterValue(t_channel);
 
  private slots:
   void slotChangeOutputPlugin(const QString& name);
